% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_lw_to_al.R
\name{apply_lw_to_al}
\alias{apply_lw_to_al}
\title{Apply a length–weight key to empirical length-at-age to get quarter WAA}
\usage{
apply_lw_to_al(
  al_emp,
  lw_fit,
  gear_mode = c("combined", "by_gear"),
  sex_mode = c("combined", "split"),
  model_length = TRUE
)
}
\arguments{
\item{al_emp}{output of build_empirical_length_at_age()}

\item{lw_fit}{output of fit_lw_key()}

\item{gear_mode}{"combined" or "by_gear"}

\item{sex_mode}{"combined" or "split"}
}
\value{
data.table with YEAR, REGION_GRP,SEASON, AGE_YRS, (GEAR2), (SEX), WAA, N_AL
}
\description{
Apply a length–weight key to empirical length-at-age to get quarter WAA
}
