% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_utils.R
\name{collapse_filters_chr}
\alias{collapse_filters_chr}
\title{Collapse character values into a SQL list (quoted)}
\usage{
collapse_filters_chr(x)
}
\arguments{
\item{x}{Character vector. NAs are dropped.}
}
\value{
A length-1 character string like "'a','b'". Returns "NULL" if empty.
}
\description{
Collapse character values into a SQL list (quoted)
}
\keyword{internal}
