% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWAA_utils.R
\name{.ewaa_agecols}
\alias{.ewaa_agecols}
\title{Identify age columns in a wide EWAA matrix}
\usage{
.ewaa_agecols(wide, maxage)
}
\description{
Identify age columns in a wide EWAA matrix
}
\keyword{internal}
