% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lw_key.R
\name{fit_lw_key}
\alias{fit_lw_key}
\title{Fit a length–weight key using fishery length–weight data}
\usage{
fit_lw_key(
  fish_data,
  trim_outliers = TRUE,
  trim_q = c(0.01, 0.99),
  min_strata_n = 1,
  allow_year = TRUE,
  allow_season = TRUE,
  allow_gear = TRUE,
  allow_region = TRUE,
  allow_area = FALSE,
  allow_sex = TRUE
)
}
\arguments{
\item{fish_data}{data.table from get_fishery_age_wt_data()}

\item{trim_outliers}{logical; if TRUE, trim extreme weights using qgam envelope on s(LENGTH)}

\item{trim_q}{length-2 quantiles for trimming}

\item{min_strata_n}{minimum N per stratum cell before retaining a factor term}

\item{allow_year, allow_quarter, allow_gear, allow_region, allow_area, allow_sex}{include those factor terms if supported}
}
\value{
list(model=mgcv::gam, kept_terms=character, data_used=data.table)
}
\description{
Fit a length–weight key using fishery length–weight data
}
