% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWAA_utils.R
\name{plot_weight_at_age_outlier}
\alias{plot_weight_at_age_outlier}
\title{Plot weight-at-age outlier diagnostics}
\usage{
plot_weight_at_age_outlier(data, ...)
}
\arguments{
\item{data}{Long-format EWAA data. If \code{outlier} column not present, it is computed.}

\item{...}{Variables to facet by (passed to facet_wrap).}
}
\value{
A cowplot object.
}
\description{
Produces length-weight and age-weight scatterplots (including log scales),
with outliers highlighted.
}
