% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_utils.R
\name{sql_read}
\alias{sql_read}
\title{Read a SQL file either from installed package extdata or from a direct path}
\usage{
sql_read(x, package = utils::packageName(), root_dir = "sql")
}
\arguments{
\item{x}{File name like "afsc/catch.sql" OR a full/relative file path.}

\item{package}{Package name to search in. Defaults to current package.}

\item{root_dir}{Subfolder inside extdata (default "sql").}
}
\value{
Character vector of SQL lines.
}
\description{
Read a SQL file either from installed package extdata or from a direct path
}
