% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_utils.R
\name{sql_reader}
\alias{sql_reader}
\title{Read a SQL file bundled with DR4SS-NP}
\usage{
sql_reader(filename)
}
\arguments{
\item{filename}{SQL file name (e.g., "LL_RPN.sql")}

\item{package}{Package name (default "DR4SS-NP")}
}
\value{
Character vector of SQL lines
}
\description{
Read a SQL file bundled with DR4SS-NP
}
