% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_utils.R
\name{sql_run}
\alias{sql_run}
\title{Run a SQL query via DBI}
\usage{
sql_run(database, query, as.is = TRUE)
}
\arguments{
\item{database}{A DBI connection.}

\item{query}{Character vector of SQL lines or a single SQL string.}

\item{as.is}{Passed to \code{DBI::dbGetQuery()}.}
}
\value{
A data.frame result from \code{DBI::dbGetQuery()}.
}
\description{
Run a SQL query via DBI
}
