% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWAA_utils.R
\name{weight_at_age_read}
\alias{weight_at_age_read}
\title{Read weight-at-age files from disk}
\usage{
weight_at_age_read(file)
}
\arguments{
\item{file}{Path to a \code{.csv} file with columns: Source, Weight_kg, Age_yrs,
Length_cm, Month, Year (Sex column is optional and will be dropped).}
}
\value{
A data frame of EWAA information.
}
\description{
Reads a \code{.csv} file containing the required EWAA columns and normalizes \code{Source}.
}
